Restructure packaging for convenience stores jp
COAND-497
